TIP-1073: Use the new WriteValueCollectionFactory
