+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -69,7 +69,7 @@ services:
     pim_catalog.event_subscriber.load_product_values:
         class: '%pim_catalog.event_subscriber.load_product_values.class%'
         arguments:
-            - '@pim_catalog.factory.value_collection'
+            - '@akeneo.pim.enrichment.factory.read.write_value_collection'
         tags:
             - { name: doctrine.event_subscriber, priority: 50 }
