+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/validators.yml
@@ -133,7 +133,7 @@ services:
         class: '%pim_catalog.validator.constraint.immutable_variant_axis_values.class%'
         arguments:
             - '@pim_catalog.family_variant.provider.entity_with_family_variant_attributes'
-            - '@pim_catalog.factory.value_collection'
+            - '@akeneo.pim.enrichment.factory.read.write_value_collection'
         tags:
             - { name: validator.constraint_validator, alias: pim_immutable_variant_axis_values_validator }
