+++ src/Akeneo/Pim/Enrichment/Component/Product/Factory/Read/WriteValueCollectionFactory.php
@@ -1,37 +1,24 @@
 <?php
 
-namespace Akeneo\Pim\Enrichment\Component\Product\Factory\Write;
+namespace Akeneo\Pim\Enrichment\Component\Product\Factory\Read;
 
-use Akeneo\Pim\Enrichment\Component\Product\Exception\InvalidAttributeException;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\EmptyValuesCleaner;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\NonExistentValuesFilter\ChainedNonExistentValuesFilterInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\NonExistentValuesFilter\OnGoingFilteredRawValues;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\TransformRawValuesCollections;
-use Akeneo\Pim\Enrichment\Component\Product\Factory\ValueFactory;
 use Akeneo\Pim\Enrichment\Component\Product\Model\WriteValueCollection;
 use Akeneo\Pim\Structure\Component\Query\PublicApi\AttributeType\GetAttributes;
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyException;
-use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
-use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
-use Psr\Log\LoggerInterface;
 
 /**
- * Create a product value collection.
- *
- * @author    Julien Janvier <j.janvier@gmail.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
+ * @author    Anael Chardan <anael.chardan@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class WriteValueCollectionFactory
 {
     /** @var ValueFactory */
     private $valueFactory;
-
-    /** @var IdentifiableObjectRepositoryInterface */
-    private $attributeRepository;
-
-    /** @var LoggerInterface */
-    private $logger;
 
     /** @var GetAttributes */
     private $getAttributeByCodes;
@@ -47,16 +34,12 @@ class WriteValueCollectionFactory
 
     public function __construct(
         ValueFactory $valueFactory,
-        IdentifiableObjectRepositoryInterface $attributeRepository,
-        LoggerInterface $logger,
         GetAttributes $getAttributeByCodes,
         ChainedNonExistentValuesFilterInterface $chainedNonExistentValuesFilter,
         EmptyValuesCleaner $emptyValuesCleaner,
         TransformRawValuesCollections $transformRawValuesCollections
     ) {
         $this->valueFactory = $valueFactory;
-        $this->attributeRepository = $attributeRepository;
-        $this->logger = $logger;
         $this->getAttributeByCodes = $getAttributeByCodes;
         $this->chainedNonExistentValuesFilter = $chainedNonExistentValuesFilter;
         $this->emptyValuesCleaner = $emptyValuesCleaner;
@@ -73,15 +56,6 @@ class WriteValueCollectionFactory
     public function createMultipleFromStorageFormat(array $rawValueCollections): array
     {
         $rawValueCollectionsIndexedByType = $this->transformRawValuesCollections->toValueCollectionsIndexedByType($rawValueCollections);
-        $valueCollections = [];
-
-        if (empty($rawValueCollectionsIndexedByType)) {
-            foreach (array_keys($rawValueCollections) as $identifier) {
-                $valueCollections[$identifier] = new WriteValueCollection([]);
-            }
-
-            return $valueCollections;
-        }
 
         $filtered = $this->chainedNonExistentValuesFilter->filterAll(
             OnGoingFilteredRawValues::fromNonFilteredValuesCollectionIndexedByType($rawValueCollectionsIndexedByType)
@@ -138,7 +112,11 @@ class WriteValueCollectionFactory
                             $localeCode = null;
                         }
 
-                        $values[] = $this->valueFactory->create($attribute, $channelCode, $localeCode, $data, true);
+                        try {
+                            //TEMPORARY
+                            $values[] = $this->valueFactory->createByCheckingData($attribute, $channelCode, $localeCode, $data);
+                        } catch (InvalidPropertyException $exception) {
+                        }
                     }
                 }
             }
