+++ src/Akeneo/Tool/Bundle/FileStorageBundle/Resources/config/repositories.yml
@@ -4,7 +4,8 @@ parameters:
 services:
     akeneo_file_storage.repository.file_info:
         class: '%akeneo_file_storage.repository.file_info.class%'
-        factory: 'doctrine.orm.entity_manager:getRepository'
-        arguments: ['%akeneo_file_storage.model.file_info.class%']
+        factory: ['@doctrine.orm.entity_manager', 'getRepository']
+        arguments:
+            - '%akeneo_file_storage.model.file_info.class%'
         tags:
             - { name: 'akeneo_repository' }
